Add files via upload
Cloud computing, with characteristics of natural information data sharing with low maintenance and better utilization of resources. In this data can be shared data in secured manner, in cloud it can be achieve secure data sharing in dynamic groups. Cloud computing offers an infinite storage space. In our scheme, secured data sharing
 
can be protected from collusion attack. In this paper the main contributions of this scheme include:


1.	The key distribution without any communication channel and the user can securely know their private key from their group manager without any certificate authority because of verification of public key of the user.

2.	This scheme can achieve fine grain access control, any user in the group can access their resources and revoked user cannot access the data in the cloud after they are rejected.

3.	This can protect collusion attack which means the revoked user cannot get original data from the cloud.

4.	Our scheme can achieve secure user revocation with the help of polynomial function.

5.	This scheme is able to achieve fine efficiency, scheme achieve fine efficiency, that is previous users need not updated they are private key when new user adds or rejected from the group.